+++ magik-squid/src/test/java/org/stevenlooman/sw/magik/parser/MagikParserTest.java
@@ -1,7 +1,12 @@
 package org.stevenlooman.sw.magik.parser;
 
 import static org.fest.assertions.Assertions.assertThat;
+
+import com.sonar.sslr.api.AstNode;
 import org.junit.Test;
+import org.stevenlooman.sw.magik.api.MagikGrammar;
+
+import java.nio.charset.Charset;
 
 public class MagikParserTest {
 
@@ -17,6 +22,22 @@ public class MagikParserTest {
     assertThat(MagikParser.parseIdentifier("abc| |")).isEqualTo("abc ");
     assertThat(MagikParser.parseIdentifier("\\|")).isEqualTo("|");
     assertThat(MagikParser.parseIdentifier("\\|a")).isEqualTo("|a");
+  }
+
+  @Test
+  public void testInvalidMagik() {
+    MagikParser parser = new MagikParser(Charset.defaultCharset());
+    String code =
+        "_block\n" +
+        "\t_loop\n" +
+        "\t_endloo\n" +
+        "_endblock\n";
+    AstNode node = parser.parse(code);
+    assertThat(node.getChildren()).hasSize(1);
+    AstNode parserErrorNode = node.getChildren().get(0);
+    assertThat(parserErrorNode.getType()).isEqualTo(MagikGrammar.PARSER_ERROR);
+    assertThat(parserErrorNode.getToken().getLine()).isEqualTo(3);
+    assertThat(parserErrorNode.getToken().getColumn()).isEqualTo(2);
   }
 
 }